refactor(runner): extract nbd sysfs helpers into shared module
Move read_nbds_max() and read_nbd_pid() from gc.rs into cmd/nbd.rs
so both gc and doctor reference a single implementation.
