+++ crates/runner/src/cmd/doctor.rs
@@ -196,22 +196,12 @@ impl Warning {
                 let idx = *device_index;
                 let original_pid = *pid;
                 tokio::task::spawn_blocking(move || {
-                    let path = format!("/sys/block/nbd{idx}/pid");
-                    match std::fs::read_to_string(&path) {
-                        Ok(content) => {
-                            let trimmed = content.trim();
-                            if trimmed.is_empty() || trimmed == "-1" || trimmed == "0" {
-                                return false; // device freed
-                            }
-                            match trimmed.parse::<u32>() {
-                                Ok(current_pid) => {
-                                    // Still orphaned if same dead PID, or new dead PID
-                                    current_pid == original_pid && !pid_exists(current_pid)
-                                }
-                                Err(_) => false,
-                            }
+                    match super::nbd::read_nbd_pid(idx) {
+                        Some(current_pid) => {
+                            // Still orphaned if same dead PID
+                            current_pid == original_pid && !pid_exists(current_pid)
                         }
-                        Err(_) => false, // pid file gone → device freed
+                        None => false, // device freed or pid cleared
                     }
                 })
                 .await
@@ -814,31 +804,16 @@ async fn detect_orphan_namespaces() -> Vec<Warning> {
     warnings
 }
 
-/// Default upper bound for NBD device indices when `/sys/module/nbd/parameters/nbds_max`
-/// is unreadable (e.g. module not loaded).
-const NBD_DEFAULT_MAX: u32 = 256;
-
 /// Scan for NBD devices whose owning process has exited without disconnecting.
 async fn detect_nbd_orphans() -> Vec<Warning> {
     let orphans = tokio::task::spawn_blocking(|| {
-        let max_devs = std::fs::read_to_string("/sys/module/nbd/parameters/nbds_max")
-            .ok()
-            .and_then(|s| s.trim().parse::<u32>().ok())
-            .unwrap_or(NBD_DEFAULT_MAX);
-
+        let max_devs = super::nbd::read_nbds_max();
         let mut found: Vec<(u32, u32)> = Vec::new();
         for i in 0..max_devs {
-            let pid_path = format!("/sys/block/nbd{i}/pid");
-            if let Ok(content) = std::fs::read_to_string(&pid_path) {
-                let trimmed = content.trim();
-                if trimmed.is_empty() || trimmed == "-1" || trimmed == "0" {
-                    continue;
-                }
-                if let Ok(pid) = trimmed.parse::<u32>()
-                    && !Path::new(&format!("/proc/{pid}")).exists()
-                {
-                    found.push((i, pid));
-                }
+            if let Some(pid) = super::nbd::read_nbd_pid(i)
+                && !Path::new(&format!("/proc/{pid}")).exists()
+            {
+                found.push((i, pid));
             }
         }
         found
